refactor: SidebarMenuCall, Added to separate the logic
